currently, souper attempts to synthesize a better computation for all (#379)
integer-typed registers, even those that have no uses. previously this
was a bit wasteful, but not problematic. however, since turning
demanded bits on by default, souper started synthesizing and replacing
every non-used return value from a function, which is just silly.

this patch causes souper to only extract registers that have at least
one use. the patch is trivial but it broke a lot of tests, which are
all fixed in obvious ways.